requirements.txt
streamlit
requests
Pillow
google-cloud-texttospeech
google-cloud-speech
google-auth
pinecone-client
datasets
tqdm
sentence-transformers
numpy
faiss-cpu
pickle5
transformers
torch
requests
beautifulsoup4
fpdf